Fix:Fixed warning with paginate on site
Fixed warning, when url page have str, float, number is more than the
number of pages, number is less than the number of pages and number
is equally zero pages. Added aginate_queryset for BaseDataMixin,
which tracks entered data on the url page and returns the appropriate
number. And added SafePaginator has been added to track the entered
data too.

Changes to be committed:
	modified:   core/views.py
	modified:   goods/mixins/goods_mixins.py
	modified:   goods/templates/goods/choice_type_goods.html
	modified:   goods/templates/goods/search_goods.html
	modified:   goods/views.py
	modified:   goods_favourite/views.py
	modified:   templates/base.html